requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cloudinary==1.26.0
coverage==5.5
Django==3.2.5
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-email-verification==0.1.0
django-js-asset==1.2.2
django-nose==1.4.7
gunicorn==20.1.0
nose==1.3.7
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6